feat: admin authorization added

+++ QuizServer/src/main/java/com/quizserver/controllers/QuizController.java
@@ -6,7 +6,10 @@ import com.quizserver.models.DTOs.commands.UserQuizAnswersCommand;
 import com.quizserver.models.DTOs.queries.QuizQuery;
 import com.quizserver.models.DTOs.queries.UpdateQuizQuery;
 import com.quizserver.models.DTOs.queries.UserQuizScoreQuery;
+import com.quizserver.services.AuthService;
 import com.quizserver.services.QuizService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -17,9 +20,11 @@ import java.util.UUID;
 public class QuizController {
 
     private final QuizService quizService;
+    private final AuthService authService;
 
-    public QuizController(final QuizService quizService) {
+    public QuizController(final QuizService quizService, final AuthService authService) {
         this.quizService = quizService;
+        this.authService = authService;
     }
 
     @GetMapping
@@ -33,27 +38,36 @@ public class QuizController {
     }
 
     @GetMapping("{quizId}/updateForm")
-    public UpdateQuizQuery getQuizByQuizIdForUpdate(@PathVariable("quizId") UUID quizId) {
-        return quizService.getQuizByQuizIdForUpdate(quizId);
+    public ResponseEntity<UpdateQuizQuery> getQuizByQuizIdForUpdate(@PathVariable("quizId") UUID quizId, @RequestHeader(value = "Authorization", defaultValue = "token") String token) {
+        System.out.println(token);
+        if(!authService.isAdmin(token)) return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        return new ResponseEntity<>(quizService.getQuizByQuizIdForUpdate(quizId), HttpStatus.OK);
     }
 
     @PostMapping
-    public void postQuiz(@RequestBody CreateQuizCommand quizCommand) {
+    public ResponseEntity<?> postQuiz(@RequestBody CreateQuizCommand quizCommand, @RequestHeader(value = "Authorization", defaultValue = "token") String token) {
+        if(!authService.isAdmin(token)) return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
         quizService.createQuiz(quizCommand);
+        return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
     @PutMapping("{quizId}")
-    public void postQuiz(@PathVariable UUID quizId, @RequestBody UpdateQuizCommand quizCommand) {
+    public ResponseEntity<?> postQuiz(@PathVariable UUID quizId, @RequestBody UpdateQuizCommand quizCommand, @RequestHeader(value = "Authorization", defaultValue = "token") String token) {
+        if(!authService.isAdmin(token)) return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
         quizService.updateQuiz(quizId, quizCommand);
+        return new ResponseEntity<>(HttpStatus.CREATED);
     }
+
     @PostMapping("{quizId}/calculateScore")
     public UserQuizScoreQuery calculateQuizScore(@PathVariable("quizId") UUID quizId, @RequestBody List<UserQuizAnswersCommand> userAnswers) {
         return quizService.calculateQuizScore(quizId, userAnswers);
     }
 
     @DeleteMapping("{quizId}")
-    public void deleteQuiz(@PathVariable UUID quizId) {
+    public ResponseEntity<?> deleteQuiz(@PathVariable UUID quizId, @RequestHeader(value = "Authorization", defaultValue = "token") String token) {
+        if(!authService.isAdmin(token)) return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
         quizService.deleteQuiz(quizId);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
 }
